Fixed requested changes
