--- src/src_V2/components/table/styles.js
@@ -1,3 +1,0 @@
-const styles = {};
-
-export default styles;
