+++ src/src_V2/modules/manager/container/sidebar/Sidebar.module.css
@@ -17,7 +17,7 @@
   left: 12px;
   width: 40px !important;
   height: 40px !important;
-  z-index: 1;
+  z-index: 11;
   position: fixed !important;
   background-color: var(--white) !important;
   transition: left 250ms cubic-bezier(0.4, 0, 0.6, 1) 0ms !important;
