+++ src/src_V2/modules/manager/user-management/UserManagementContent/styles.module.css
@@ -1,4 +1,4 @@
-@import '../../../styles/colors.css';
+@import '../../../../styles/colors.css';
 
 .content .table {
   table-layout: fixed;
