+++ src/src_V2/modules/manager/user-management/utils/table-data-generator.js
@@ -6,7 +6,7 @@ const userManagementTableDataGenerator = (entries) => ({
     createTableColumn('User Position', 'User Position'),
     createTableColumn('Email', 'Email'),
     createTableColumn('Phone Number', 'Phone Number'),
-    createTableColumn('Status', 'Status'),
+    createTableColumn('Status', 'Status', '', { width: 160 }),
     createTableColumn('editBtn', '', '', { width: 66 }),
   ],
   rows: entries.map((data) =>
@@ -17,7 +17,7 @@ const userManagementTableDataGenerator = (entries) => ({
         createTableColumn('position', data.position),
         createTableColumn('email', data.email),
         createTableColumn('phone', data.phone),
-        createTableColumn('status', data.status),
+        createTableColumn('status', data.status, '', { width: 160 }),
         createTableColumn('editBtn', data.editBtn, '', { width: 66 }),
       ],
       data
