+++ src/src_V2/utils/constants.js
@@ -10,7 +10,7 @@ export const ERROR_MESSAGES = {
   integer: 'The filed must be an integer',
   phone: 'The field is not a valid phone number',
   password: 'All validation levels must be passed',
-  repeatPassword: 'Passwords doesn\'t match',
+  repeatPassword: 'Passwords don\'t match',
   maxNumberField: (max) => `The value must be at most ${max}`,
   maxTextField: (max) => `The filed must be at most ${max} characters`,
 };
