+++ src/src_V2/utils/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useHoverTooltip } from './useHoverTooltip';
 export { default as useTimer } from './useTimer';
+export { default as useToggle } from './useToggle';
 export { default as useTableDataGenerator } from './use-table-data-generator';
